change double column of amount to decimal

+++ database/migrations/2024_06_22_033334_add_amount_to_otcs_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('otcs', function (Blueprint $table) {
-            $table->double('amount')->after('name');
+            $table->decimal('amount', 8, 2)->after('name');
         });
     }
